#4303 flip view tabs
